+++ Src/Library/Config/ErrorOptions.cs
@@ -15,10 +15,13 @@ public sealed class ErrorOptions
     public int StatusCode { internal get; set; } = 400;
 
     /// <summary>
-    /// if this property is set, a <see cref="IProducesResponseTypeMetadata"/> will automatically be added to any endpoints that has a <see cref="Validator{TRequest}"/> associated with it.
-    /// if you're not specifying your own <see cref="ResponseBuilder"/>, you'd typically be setting this to <c>typeOf(<see cref="ErrorResponse"/>)</c>
+    /// if this property is not null, a <see cref="IProducesResponseTypeMetadata"/> will automatically be added to endpoints that has a <see cref="Validator{TRequest}"/> associated with it.
+    /// if you're specifying your own <see cref="ResponseBuilder"/>, don't forget to set this property to the correct type of the error response dto that your error response builder will be returning.
+    /// <para>
+    /// TIP: set this to null if you'd like to disable the auto adding of produces 400 metadata to endpoints even if they have validators associated with them.
+    /// </para>
     /// </summary>
-    public Type? ProducesMetadataType { internal get; set; }
+    public Type? ProducesMetadataType { internal get; set; } = typeof(ErrorResponse);
 
     /// <summary>
     /// the general errors field name. this is the field name used for general errors when AddError() method is called without specifying a request dto property.
@@ -30,6 +33,9 @@ public sealed class ErrorOptions
     /// set it to any func that returns an object that can be serialized to json.
     /// this function will be run everytime an error response needs to be sent to the client.
     /// the arguments for the func will be a list of validation failures, the http context and an http status code.
+    /// <para>
+    /// HINT: if changing the default, make sure to also set <see cref="ProducesMetadataType"/> to the correct type of the error response dto.
+    /// </para>
     /// </summary>
     public Func<List<ValidationFailure>, HttpContext, int, object> ResponseBuilder { internal get; set; }
         = (failures, _, statusCode) => new ErrorResponse(failures, statusCode);